+++ src/com/kaylerrenslow/armaplugin/lang/sqf/psi/SQFSwitchStatement.java
@@ -1,0 +1,8 @@
+package com.kaylerrenslow.armaplugin.lang.sqf.psi;
+
+/**
+ * @author Kayler
+ * @since 09/18/2017
+ */
+public class SQFSwitchStatement implements SQFControlStructure {
+}